requirements.txt
torch>=2.0
torchvision>=0.15
numpy>=1.24
pandas>=1.5
matplotlib>=3.7
scikit-learn>=1.2
opencv-python>=4.7
Pillow>=9.4
seaborn>=0.12
jupyter>=1.0
flask>=2.3